llm-evaluation/llm-prompt-versioning-and-evaluation-streamlit-app/requirements.txt
streamlit>=1.38.0
mlflow>=3.3
pydantic>=2.0
groq>=0.11.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
